Version 3.0 Release Candidatde
Here is what's new in short:
- Some of the core code components have been moved to their own swift packages, see the packages dependancies of the xcode project to learn more.
- Due to several changes in how the app works the EFI Partition Mounter tool is no longer buildable as a standalone tool.
- New design for some UI Elements.
- Added icons for the app's buttons.
- New design for EFI Partition Mounter.
- Replaced many old managers and system with new ones.
- Unification and re-organisation of many of the app's core components.
- Fixes of long-standing bugs.
- Fixed of text errors and spelling mistakes.

There are many other changes, but they are so many that it isn't really worth talking about them, if you are curious just check out the source code.

This commit also signals a new for this project to be maintained, over time parts of the code will be ported to their own libraries, also commits on this repo will be more active since it will be used from now on as the main development repository.